Disable ERROR_ON_CLONE
This was pulled from the devstack-gate config
but it really doesn't make sense for us at the
moment as we intend to clone all repositories.

We might change this if we create an image
that preseeds everything except Nova or uses
shared volumes?

For now, though, this setting is broken.