Download apt packages that are NOPRIME
The NOPRIME packages aren't installed with install_prereqs.sh,
they're only supposed to be installed if and when necessary during
the stack.sh run. That's a good thing.

The bad thing is that our images are thus dependent on the network
and these files aren't being cached. This change downloads those
deb packages during the Docker build, speeding up the 'docker run'.