Fix dbus for libvirt based drivers
I encountered errors starting libvirt if dbus was
not running, although it's not a required package.

This forces dbus to be installed and creates a symlink
to fix a bug that exists in its package where its init script
attempts to use the wrong path to the dbus-daemon executable.

This only affects running devstack with libvirt based drivers
(this does not affect docker-in-docker installs)